// src/core/model/types.ts
export interface SystemMessage {
  role: 'system'
  content: string
  name?: string
}

export interface UserMessage {
  role: 'user'
  content: string
  name?: string
}

export interface AssistantMessage {
  role: 'assistant'
  content: string | null
  name?: string
  tool_calls?: ChatCompletionTool[]
  prefix?: boolean
  reasoning_content?: string
}

export interface ToolMessage {
  role: 'tool'
  content: string
  tool_call_id: string
  name?: string
}

export type ChatMessage = SystemMessage | UserMessage | AssistantMessage | ToolMessage
// ...
export interface Usage {
  /**
   * Number of tokens in the generated completion.
   */
  completion_tokens: number
  /**
   * Number of tokens in the prompt.
   *
   * It equals prompt_cache_hit_tokens + prompt_cache_miss_tokens.
   */
  prompt_tokens: number
  /**
   * Number of tokens in the prompt that hits the context cache.
   */
  prompt_cache_hit_tokens: number
  /**
   * Number of tokens in the prompt that misses the context cache.
   */
  prompt_cache_miss_tokens: number
  /**
   * Total number of tokens used in the request (prompt + completion).
   */
  total_tokens: number
  /**
   * Breakdown of tokens used in a completion.
   */
  completion_tokens_details: {
    /**
     * Tokens generated by the model for reasoning.
     */
    reasoning_tokens: number
  }
}
// ...
export interface ChatCompletionChunkDelta {
  role?: 'assistant'
  content?: string
  reasoning_content?: string
  tool_calls?: {
    index: number
    id?: string
    type?: 'function'
    function?: {
      name?: string
      arguments?: string
    }
  }[]
}